requirements.txt
asgiref==3.3.1
Django==3.1.4
django-extensions==3.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
importlib-metadata==3.3.0
Markdown==3.3.3
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
